admin: add CoverImageTrait
This trait will allow reuse of the cover image code from
LiveItemFormModel for the upcoming EpisodeFormModel class.
